Reorganize project for NPM dependencies
Remove dependencies included as file in the repo in favor of getting
these dependencies from NPM

 * ace (We use "brace")
 * diff_match_patch

We remove the local jQuery copy and will use a CDN version for the
demos.

Remove gruntfile (I like webpack more and we hardly need any of them
here), also remove bower.. doesn't seem very sensible for this endeavor.

move source to "src".

Set up brace as peerDependency to avoid the same double-draw-in-mess we
do this for in the first place.